backend/security-tests/requirements.txt
# Python dependencies for security evaluation framework

# HTTP client for API testing
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Testing framework
pytest>=7.4.3
pytest-asyncio>=0.21.1

# JSONL file parsing
jsonlines>=4.0.0

# Date/time utilities
python-dateutil>=2.8.2

# Colorized terminal output
colorama>=0.4.6

# Progress bars
tqdm>=4.66.1

# Data validation
pydantic>=2.5.0

# JSON schema validation
jsonschema>=4.20.0

# Retry logic for HTTP requests
tenacity>=8.2.3

# Statistical analysis
numpy>=1.26.2

# Report generation
jinja2>=3.1.2

# Markdown to HTML (for reports)
markdown>=3.5.1
